pyproject.toml
[project]
name = "brother-label"
version = "1.1a0"
description = "Python package to talk to Brother QL label printers"
readme = "README.md"
authors = [
    { name = "Philipp Klaus", email = "philipp.l.klaus@web.de" },
    { name = "Dean Gardiner", email = "me@dgardiner.net" }
]
license = {text = "GPL"}
dependencies = [
    "click",
    "future",
    "packbits",
    "pillow>=3.3.0",
    "pyusb",
    "attrs",
    "typing;python_version<'3.5'",
    "enum34;python_version<'3.4'",
]
keywords = [
    "Brother",
    "QL-500",
    "QL-550",
    "QL-560",
    "QL-570",
    "QL-600",
    "QL-700",
    "QL-710W",
    "QL-720NW",
    "QL-800",
    "QL-810W",
    "QL-820NWB",
    "QL-1050",
    "QL-1060N",
    "QL-1100",
    "QL-1100NWB",
    "QL-1115NWB",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Operating System :: OS Independent",
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering :: Visualization",
    "Topic :: System :: Hardware :: Hardware Drivers",
]

[project.scripts]
brother_ql = "brother_ql.cli:cli"

[project.urls]
repository = "https://github.com/fuzeman/brother-label"
